login test: assertion on fetchSpy
