+++ packages/cfa-template/template/frontend/__mocks__/sessionStorage.js
@@ -1,0 +1,22 @@
+const sessionStorageMock = (() => {
+    let store = {};
+
+    return {
+        getItem(key) {
+            return store[key] || null;
+        },
+        setItem(key, value) {
+            store[key] = value.toString();
+        },
+        removeItem(key) {
+            delete store[key];
+        },
+        clear() {
+            store = {};
+        },
+    };
+})();
+
+Object.defineProperty(window, "sessionStorage", {
+    value: sessionStorageMock,
+});
